feat: dynamic tunnel offset, 10yr history, date fixes, methodology
Task 1: Replace static +4°C tunnel offset with dynamic radiation-based
model. Fetches is_day + shortwave_radiation from Open-Meteo hourly.
Night = 0°C offset, sunny noon = up to +15°C. Formula:
offset = min(15, radiation * 15/800).

Task 2: Extend historical weather from 3 years to 10+ years default
for robust P10/P50/P90 climate percentiles.

Task 3: Verified variety-specific GDH thresholds already working
(section override → variety fallback → null).

Task 4: Fix Polish date formatting from nominative ("maj") to
genitive ("maja") using month:'long' in toLocaleDateString.

Task 5: Complete rewrite of methodology info box with dynamic offset
model, radiation formula, updated examples, ECMWF line description,
per-variety threshold explanation.

Also: Fix plantingDate not persisting (ISO datetime→YYYY-MM-DD format
conversion bug when loading into date input).

+++ src/app/dashboard/plantation/gdh-matrix.tsx
@@ -343,8 +343,8 @@ export default function GDHMatrix() {
               return fruitCell ? fruitCell.weekStart : null
             }).filter(Boolean) as string[]
 
-            const earliest = (dates: string[]) => dates.length ? new Date(dates.sort()[0]).toLocaleDateString('pl-PL', { day: 'numeric', month: 'short' }) : '—'
-            const latest = (dates: string[]) => dates.length ? new Date(dates.sort().pop()!).toLocaleDateString('pl-PL', { day: 'numeric', month: 'short' }) : '—'
+            const earliest = (dates: string[]) => dates.length ? new Date(dates.sort()[0]).toLocaleDateString('pl-PL', { day: 'numeric', month: 'long' }) : '—'
+            const latest = (dates: string[]) => dates.length ? new Date(dates.sort().pop()!).toLocaleDateString('pl-PL', { day: 'numeric', month: 'long' }) : '—'
 
             return (
               <>
